project/python/requirements.txt
torch~=1.8.1
torchvision~=0.9.1
Pillow~=8.2.0
fire~=0.4.0
numpy~=1.20.2
tqdm~=4.61.0